+++ packages/yucca-sdk/orchestration-ui/src/lib/components/backups/BackupsList.svelte
@@ -4,6 +4,7 @@
     useRepositories,
     useRepositoryEventHandler,
   } from "$lib/services/repository.service";
+  import { getProvider } from "$lib/providers";
   import { Button, HStack, modalManager, Stack } from "@immich/ui";
   import StackList from "../ui/StackList.svelte";
   import StackListPlaceholder from "../ui/StackListPlaceholder.svelte";
@@ -13,11 +14,12 @@
   import CreateRepositoryModal from "./dialogs/CreateRepositoryModal.svelte";
 
   type Props = {
-    local?: boolean;
     initialData?: RepositoryListResponseDto;
   };
 
-  const { local, initialData }: Props = $props();
+  const { initialData }: Props = $props();
+
+  const local = getProvider().api === "orchestrator";
 
   // svelte-ignore state_referenced_locally
   const query = useRepositories(initialData?.repositories);
@@ -46,7 +48,7 @@
 
         <Suspense {query}>
           {#each localRepositories as repository (repository.id)}
-            <BackupItem {repository} {local} />
+            <BackupItem {repository} />
           {/each}
 
           {#if localRepositories.length === 0}